# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Homestay {
  id: number;
  name: string;
  host: string;
  village: string;
  region: string;
  state: string;
  /** INR per night */
  pricePerNight: number;
  rating: number;
  reviewCount: number;
  imageUrl: string;
  activities: string[];
  tags: string[];
}

export type HomestayDetail = Homestay & {
  description: string;
  maxGuests: number;
  bedrooms: number;
  /** 0–100 host readiness score */
  readinessScore: number;
};

export interface Booking {
  id: number;
  homestayId: number;
  guestName: string;
  guestEmail?: string;
  checkInDate: string;
  guests: number;
  createdAt: string;
}

export interface CreateBookingBody {
  homestayId: number;
  /** @minLength 2 */
  guestName: string;
  guestEmail?: string;
  checkInDate: string;
  /**
   * @minimum 1
   * @maximum 12
   */
  guests: number;
}

export interface ChecklistItem {
  key: string;
  label: string;
  checked: boolean;
}

export interface ChecklistResult {
  homestayId: number;
  items: ChecklistItem[];
  /** 0–100 readiness score */
  score: number;
  updatedAt: string;
}

export type SaveChecklistBodyItemsItem = {
  key: string;
  checked: boolean;
};

export interface SaveChecklistBody {
  items: SaveChecklistBodyItemsItem[];
}

export type InsightsSummaryTopRegionsItem = {
  region: string;
  count: number;
};

export interface InsightsSummary {
  totalHomestays: number;
  totalVillages: number;
  totalBookings: number;
  averageReadiness: number;
  topRegions: InsightsSummaryTopRegionsItem[];
}

export interface RecentBooking {
  id: number;
  guestName: string;
  checkInDate: string;
  createdAt: string;
  homestayId: number;
  homestayName: string;
  village: string;
}

export interface TopHomestay {
  id: number;
  name: string;
  village: string;
  region: string;
  readinessScore: number;
  imageUrl: string;
}

export type ListHomestaysParams = {
  region?: string;
  /**
   * Free text search across name, village, host
   */
  q?: string;
};

export type GetHomestayAvailability200 = {
  bookedDates: string[];
};

export type ListBookingsParams = {
  homestayId?: number;
};

export type GetRecentBookingsParams = {
  limit?: number;
};

export type GetTopHomestaysParams = {
  limit?: number;
};