# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List homestays
 */
export const ListHomestaysQueryParams = zod.object({
  region: zod.coerce.string().optional(),
  q: zod.coerce
    .string()
    .optional()
    .describe("Free text search across name, village, host"),
});

export const ListHomestaysResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  host: zod.string(),
  village: zod.string(),
  region: zod.string(),
  state: zod.string(),
  pricePerNight: zod.number().describe("INR per night"),
  rating: zod.number(),
  reviewCount: zod.number(),
  imageUrl: zod.string(),
  activities: zod.array(zod.string()),
  tags: zod.array(zod.string()),
});
export const ListHomestaysResponse = zod.array(ListHomestaysResponseItem);

/**
 * @summary Get a homestay by id
 */
export const GetHomestayParams = zod.object({
  id: zod.coerce.number(),
});

export const GetHomestayResponse = zod
  .object({
    id: zod.number(),
    name: zod.string(),
    host: zod.string(),
    village: zod.string(),
    region: zod.string(),
    state: zod.string(),
    pricePerNight: zod.number().describe("INR per night"),
    rating: zod.number(),
    reviewCount: zod.number(),
    imageUrl: zod.string(),
    activities: zod.array(zod.string()),
    tags: zod.array(zod.string()),
  })
  .and(
    zod.object({
      description: zod.string(),
      maxGuests: zod.number(),
      bedrooms: zod.number(),
      readinessScore: zod.number().describe("0–100 host readiness score"),
    }),
  );

/**
 * @summary Get already-booked dates for a homestay (next 90 days)
 */
export const GetHomestayAvailabilityParams = zod.object({
  id: zod.coerce.number(),
});

export const GetHomestayAvailabilityResponse = zod.object({
  bookedDates: zod.array(zod.coerce.date()),
});

/**
 * @summary List recent bookings
 */
export const ListBookingsQueryParams = zod.object({
  homestayId: zod.coerce.number().optional(),
});

export const ListBookingsResponseItem = zod.object({
  id: zod.number(),
  homestayId: zod.number(),
  guestName: zod.string(),
  guestEmail: zod.string().optional(),
  checkInDate: zod.coerce.date(),
  guests: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListBookingsResponse = zod.array(ListBookingsResponseItem);

/**
 * @summary Create a booking
 */
export const createBookingBodyGuestNameMin = 2;

export const createBookingBodyGuestsMax = 12;

export const CreateBookingBody = zod.object({
  homestayId: zod.number(),
  guestName: zod.string().min(createBookingBodyGuestNameMin),
  guestEmail: zod.string().optional(),
  checkInDate: zod.coerce.date(),
  guests: zod.number().min(1).max(createBookingBodyGuestsMax),
});

/**
 * @summary Get latest readiness checklist for a homestay
 */
export const GetHomestayChecklistParams = zod.object({
  id: zod.coerce.number(),
});

export const GetHomestayChecklistResponse = zod.object({
  homestayId: zod.number(),
  items: zod.array(
    zod.object({
      key: zod.string(),
      label: zod.string(),
      checked: zod.boolean(),
    }),
  ),
  score: zod.number().describe("0–100 readiness score"),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Save readiness checklist for a homestay
 */
export const SaveHomestayChecklistParams = zod.object({
  id: zod.coerce.number(),
});

export const SaveHomestayChecklistBody = zod.object({
  items: zod.array(
    zod.object({
      key: zod.string(),
      checked: zod.boolean(),
    }),
  ),
});

export const SaveHomestayChecklistResponse = zod.object({
  homestayId: zod.number(),
  items: zod.array(
    zod.object({
      key: zod.string(),
      label: zod.string(),
      checked: zod.boolean(),
    }),
  ),
  score: zod.number().describe("0–100 readiness score"),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Aggregated stats for the landing experience
 */
export const GetInsightsSummaryResponse = zod.object({
  totalHomestays: zod.number(),
  totalVillages: zod.number(),
  totalBookings: zod.number(),
  averageReadiness: zod.number(),
  topRegions: zod.array(
    zod.object({
      region: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Recent booking activity feed
 */
export const getRecentBookingsQueryLimitDefault = 8;

export const GetRecentBookingsQueryParams = zod.object({
  limit: zod.coerce.number().default(getRecentBookingsQueryLimitDefault),
});

export const GetRecentBookingsResponseItem = zod.object({
  id: zod.number(),
  guestName: zod.string(),
  checkInDate: zod.coerce.date(),
  createdAt: zod.coerce.date(),
  homestayId: zod.number(),
  homestayName: zod.string(),
  village: zod.string(),
});
export const GetRecentBookingsResponse = zod.array(
  GetRecentBookingsResponseItem,
);

/**
 * @summary Top homestays by readiness score
 */
export const getTopHomestaysQueryLimitDefault = 5;

export const GetTopHomestaysQueryParams = zod.object({
  limit: zod.coerce.number().default(getTopHomestaysQueryLimitDefault),
});

export const GetTopHomestaysResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  village: zod.string(),
  region: zod.string(),
  readinessScore: zod.number(),
  imageUrl: zod.string(),
});
export const GetTopHomestaysResponse = zod.array(GetTopHomestaysResponseItem);