# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  homestayId: number;
  guestName: string;
  guestEmail?: string;
  checkInDate: Date;
  guests: number;
  createdAt: Date;
}