# lib/api-zod/src/generated/types/checklistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */

export interface ChecklistItem {
  key: string;
  label: string;
  checked: boolean;
}