# lib/api-zod/src/generated/types/checklistResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */
import type { ChecklistItem } from "./checklistItem";

export interface ChecklistResult {
  homestayId: number;
  items: ChecklistItem[];
  /** 0–100 readiness score */
  score: number;
  updatedAt: Date;
}