# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  homestayId: number;
  /** @minLength 2 */
  guestName: string;
  guestEmail?: string;
  checkInDate: Date;
  /**
   * @minimum 1
   * @maximum 12
   */
  guests: number;
}