# lib/api-zod/src/generated/types/homestay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */

export interface Homestay {
  id: number;
  name: string;
  host: string;
  village: string;
  region: string;
  state: string;
  /** INR per night */
  pricePerNight: number;
  rating: number;
  reviewCount: number;
  imageUrl: string;
  activities: string[];
  tags: string[];
}