# lib/api-zod/src/generated/types/homestayDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */
import type { Homestay } from "./homestay";

export type HomestayDetail = Homestay & {
  description: string;
  maxGuests: number;
  bedrooms: number;
  /** 0–100 host readiness score */
  readinessScore: number;
};