# lib/api-zod/src/generated/types/insightsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */
import type { InsightsSummaryTopRegionsItem } from "./insightsSummaryTopRegionsItem";

export interface InsightsSummary {
  totalHomestays: number;
  totalVillages: number;
  totalBookings: number;
  averageReadiness: number;
  topRegions: InsightsSummaryTopRegionsItem[];
}