# lib/api-zod/src/generated/types/topHomestay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grama Vasathi API — rural homestays, bookings, and host readiness checklist
 * OpenAPI spec version: 0.1.0
 */

export interface TopHomestay {
  id: number;
  name: string;
  village: string;
  region: string;
  readinessScore: number;
  imageUrl: string;
}